Initial commit
Create a new Maven project with a simple IndexReporter class. When run as an executable jar, It takes a -p command line argument that is the path to the hazelcast.xml configuration file for the Metacat instance.  It then calls IndexReporter.report() and outputs a formatted JSON string with the size of the index queue and hash of the object identifiers and their positions in the queue.  Users must have read access to the hazelcast.xml file.